requirements.txt
flask==3.1.0
gunicorn==23.0.0
requests==2.32.3
python-dotenv==1.1.0
numpy>=1.24
pandas>=2.0
schedule==1.2.2
